src/main/java/io/github/manishdait/jplotlib/data/util/DataDiff.java
/*
 * MIT License
 * 
 * Copyright (c) 2023 Manish Dait
 * 
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 * 
 * The above copyright notice and this permission notice shall be included in
 * all copies or substantial portions of the Software.
 * 
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 * SOFTWARE.
 */

package io.github.manishdait.jplotlib.data.util;

/**
 * The DataDiff class provides utility methods to calculate data differences.
 *
 */

public class DataDiff {

  /**
   * Calculate the value difference based on the given points and length.
   *
   * @param points The array of points.
   * @param length The length of the data.
   * @return The value difference.
   */
  public static int getValDiff(double[] points, int length) {
    double max = 1;
      for(double j: points) {
        if(j > max) {
          max = j;
        }
      }

    
    
    int valDiff =  (int) max / length;
    if(max % points.length != 0) {
      return ++valDiff;
    }
    return valDiff;
  }
    
}
